assignment5/doctor-appointment-api/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
python-jose
passlib[bcrypt]
pydantic[email]
python-dotenv
pytest
httpx
